Allow using global types as return types
Currently referencing a global type as return type results in a generated
Java method which returns type "String". Change the code so it handles
this, just the code that handles the body (Java method parameter) does.
Also add a separate test for this to ensure that this doesn't break in the
future.