allow global type as return value (#11)
* Only test for the request body in the existing "global-type" test

This renames the existing "global-type" testcase to "global-type-body" as
the current test only ensures that a global type can be passed as body.
This makes the "User" schema in the global-type-body testcase obsolete,
which is why it's being removed.
A new test which ensures that a global-type can be used as return type
will be added separately.

* Allow using global types as return types

Currently referencing a global type as return type results in a generated
Java method which returns type "String". Change the code so it handles
this, just the code that handles the body (Java method parameter) does.
Also add a separate test for this to ensure that this doesn't break in the
future.