Add option for using BigDecimal for number type
We are using generated clients to transfer monetary data for fields that
are declared as type:number. Before this commit Double would be
generated for number types which was not adequate for transfering
monetary data.  This commit exposes the option provided by
jsonschema2pojo to generate BigDecimals for numbers.

NOTE: raml-spec does not define arbitrary precision numbers but only
provides the option to select from fixed precision formats. Despite this
we think that using BigDecimals on clients is reasonable to represent &
transfer such information. Their use is supported by various tools such
as serializers/deserializers eg jackson and code generators for other
specification eg openapi-generator-maven-plugin.